% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swell.R
\name{calculateAUC}
\alias{calculateAUC}
\title{Calculate AUC}
\usage{
calculateAUC(x, y)
}
\arguments{
\item{x}{Vector with numeric values containing all the x values (timepoints)}

\item{y}{Vector with numeric values containing all the y values, corresponding to x.}
}
\value{
The calculated area under the curve
}
\description{
This function calculates the Area Under the Curve (AUC) of x and y values.
It expects and numeric vector for the x values and a numeric vector for the corresponding y values.
More information on the AUC please check: # https://medium.com/geek-physics/3-ways-to-find-the-area-under-a-curve-and-why-you-would-want-to-do-that-949434a727d3
}
